Update README.md
This commit updates the README.md file with relevant information for me as a new software engineering student learning at ALX. The README now provides an overview of my background, learning journey, and goals. It also includes a link to my social media handles, making it easier for others to follow my progress and projects.